+++ src/data/disks.rs
@@ -11,13 +11,30 @@ pub struct Disk {
 }
 
 impl Disk {
+    pub fn new(
+        name: Option<&str>,
+        mount_point: &Path,
+        is_removable: bool,
+        canonicalize_paths: bool,
+    ) -> Self {
+        Self {
+            mount_point: canonicalize(mount_point, canonicalize_paths),
+            display_name: gen_display_name(
+                name.unwrap_or_default(),
+                mount_point.to_str().unwrap_or_default(),
+            ),
+            is_removable,
+        }
+    }
+
     /// Create a new Disk object based on the data of a `sysinfo::Disk`.
     pub fn from_sysinfo_disk(disk: &sysinfo::Disk, canonicalize_paths: bool) -> Self {
-        Self {
-            mount_point: Self::canonicalize(disk.mount_point(), canonicalize_paths),
-            display_name: gen_display_name(disk),
-            is_removable: disk.is_removable(),
-        }
+        Self::new(
+            disk.name().to_str(),
+            disk.mount_point(),
+            disk.is_removable(),
+            canonicalize_paths,
+        )
     }
 
     /// Returns the mount point of the disk
@@ -33,16 +50,6 @@ impl Disk {
     pub const fn is_removable(&self) -> bool {
         self.is_removable
     }
-
-    /// Canonicalizes the given path.
-    /// Returns the input path in case of an error.
-    fn canonicalize(path: &Path, canonicalize: bool) -> PathBuf {
-        if canonicalize {
-            dunce::canonicalize(path).unwrap_or_else(|_| path.to_path_buf())
-        } else {
-            path.to_path_buf()
-        }
-    }
 }
 
 /// Wrapper above the `sysinfo::Disks` struct
@@ -54,14 +61,9 @@ pub struct Disks {
 impl Disks {
     /// Creates a new Disks object with a refreshed list of the system disks.
     pub fn new_with_refreshed_list(canonicalize_paths: bool) -> Self {
-        let disks = sysinfo::Disks::new_with_refreshed_list();
-
-        let mut result: Vec<Disk> = Vec::new();
-        for disk in &disks {
-            result.push(Disk::from_sysinfo_disk(disk, canonicalize_paths));
+        Self {
+            disks: load_disks(canonicalize_paths),
         }
-
-        Self { disks: result }
     }
 
     /// Very simple wrapper method of the disks `.iter()` method.
@@ -71,18 +73,19 @@ impl Disks {
     }
 }
 
-#[cfg(windows)]
-fn gen_display_name(disk: &sysinfo::Disk) -> String {
-    // TODO: Get display name of the devices.
-    // Currently on Windows it returns an empty string for the C:\\ drive.
+/// Canonicalizes the given path.
+/// Returns the input path in case of an error.
+fn canonicalize(path: &Path, canonicalize: bool) -> PathBuf {
+    if canonicalize {
+        dunce::canonicalize(path).unwrap_or_else(|_| path.to_path_buf())
+    } else {
+        path.to_path_buf()
+    }
+}
 
-    let mut name = disk.name().to_str().unwrap_or_default().to_string();
-    let mount_point = disk
-        .mount_point()
-        .to_str()
-        .unwrap_or_default()
-        .to_string()
-        .replace('\\', "");
+#[cfg(windows)]
+fn gen_display_name(name: &str, mount_point: &str) -> String {
+    let mount_point = mount_point.replace('\\', "");
 
     // Try using the mount point as the display name if the specified name
     // from sysinfo::Disk is empty or contains invalid characters
@@ -90,20 +93,60 @@ fn gen_display_name(disk: &sysinfo::Disk) -> String {
         return mount_point;
     }
 
-    name.push_str(format!(" ({mount_point})").as_str());
-
-    name
+    format!("{name} ({mount_point})")
 }
 
 #[cfg(not(windows))]
-fn gen_display_name(disk: &sysinfo::Disk) -> String {
-    let name = disk.name().to_str().unwrap_or_default().to_string();
-
+fn gen_display_name(name: &str, mount_point: &str) -> String {
     // Try using the mount point as the display name if the specified name
     // from sysinfo::Disk is empty or contains invalid characters
     if name.is_empty() {
-        return disk.mount_point().to_str().unwrap_or_default().to_string();
+        return mount_point.to_string();
     }
 
-    name
+    name.to_string()
+}
+
+#[cfg(windows)]
+fn load_disks(canonicalize_paths: bool) -> Vec<Disk> {
+    let mut disks: Vec<Disk> = sysinfo::Disks::new_with_refreshed_list()
+        .iter()
+        .map(|d| Disk::from_sysinfo_disk(d, canonicalize_paths))
+        .collect();
+
+    // `sysinfo::Disks` currently do not include mapped network drives on Windows.
+    // We will load all other available drives using the Windows API.
+    // However, the sysinfo disks have priority, we are just adding to the list.
+    #[allow(unsafe_code)]
+    let mut drives = unsafe { GetLogicalDrives() };
+    let mut letter = b'A';
+
+    while drives > 0 {
+        if drives & 1 != 0 {
+            let path = PathBuf::from(format!("{}:\\", letter as char));
+            let mount_point = canonicalize(&path, canonicalize_paths);
+
+            if !disks.iter().any(|d| d.mount_point == mount_point) {
+                disks.push(Disk::new(None, &path, false, canonicalize_paths));
+            }
+        }
+
+        drives >>= 1;
+        letter += 1;
+    }
+
+    disks
+}
+
+#[cfg(windows)]
+extern "C" {
+    pub fn GetLogicalDrives() -> u32;
+}
+
+#[cfg(not(windows))]
+fn load_disks(canonicalize_paths: bool) -> Vec<Disk> {
+    sysinfo::Disks::new_with_refreshed_list()
+        .iter()
+        .map(|d| Disk::from_sysinfo_disk(d, canonicalize_paths))
+        .collect()
 }